Fix type declarations in io

+++ treeshr/treeshrp.h
@@ -793,12 +793,12 @@ extern int MDS_IO_SOCKET(int fd);
 extern int MDS_IO_FD(int fd);
 #ifdef HAVE_WINDOWS_H
 typedef int mode_t;
-typedef int ssize_t;
+typedef size_t ssize_t;
 #endif
 extern int MDS_IO_OPEN(char *filename, int options, mode_t mode);
 extern int MDS_IO_CLOSE(int fd);
 extern _int64 MDS_IO_LSEEK(int fd, _int64 offset, int whence);
-extern int MDS_IO_WRITE(int fd, void *buff, size_t count);
+extern ssize_t MDS_IO_WRITE(int fd, void *buff, size_t count);
 extern ssize_t MDS_IO_READ(int fd, void *buff, size_t count);
 extern int MDS_IO_LOCK(int fd, _int64 offset, int size, int mode, int *deleted);
 extern int MDS_IO_EXISTS(char *filename);
